changed testing focus
focus on TestClients instead of their stores, creating a SyncMultipleEngine for them
